% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nisra_colours}
\alias{nisra_colours}
\alias{old_colours}
\title{NISRA colour palettes}
\format{
\subsection{\code{nisra_colours}}{

\describe{
\item{name}{Colour name}
\item{hex}{Hex code for colour}
\item{order}{Order to use colours in}
\item{text_overlay}{Colour for text overlayed on coloured background}
\item{use_for_text}{Can this colour be used for text?}
\item{type}{"main" colour palette or "supplementary" additional colours}
}
}

An object of class \code{character} of length 3.
}
\usage{
nisra_colours

old_colours
}
\description{
An accessible colour palette for use in NISRA publications.
}
\keyword{datasets}
