% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nisra_palette.R
\name{nisra_palette}
\alias{nisra_palette}
\title{NISRA colour palettes}
\usage{
nisra_palette(reverse = FALSE)
}
\arguments{
\item{reverse}{reverse colour ordering, default is \code{FALSE}}
}
\value{
A function that takes a number (\code{n}) and returns a colour palette
with \code{n} colours
}
\description{
NISRA colour palettes
}
