% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nisracolours-package.R
\docType{package}
\name{nisracolours-package}
\alias{nisracolours}
\alias{nisracolours-package}
\title{nisracolours: Accessible Chart Colours for NISRA}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mark Paulin \email{mark.paulin@nisra.gov.uk}

}
\keyword{internal}
