% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_discrete.R
\name{scale_colour_discrete_nisra}
\alias{scale_colour_discrete_nisra}
\alias{scale_fill_discrete_nisra}
\title{Discrete NISRA colour scales}
\usage{
scale_colour_discrete_nisra(reverse = FALSE, ...)

scale_fill_discrete_nisra(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Reverse the palette, default \code{FALSE}}

\item{...}{Additional arguments passed on to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}}
}
\description{
Discrete NISRA colour scales
}
\examples{
library(ggplot2)
ggplot(mpg, aes(cty, colour = factor(drv), fill = factor(drv))) +
  geom_density(alpha = 0.2) +
  scale_fill_discrete_nisra() +
  scale_colour_discrete_nisra()
}
